Adjust bela::pe
