+++ src/belawin/pe/string.cc
@@ -1,8 +1,7 @@
-//
-#include <bela/pe.hpp>
+///
+#include "internal.hpp"
 
 namespace bela::pe {
-
 void StringTable::MoveFrom(StringTable &&other) {
   if (data != nullptr) {
     HeapFree(GetProcessHeap(), 0, data);
@@ -71,23 +70,5 @@ bool readStringTable(FileHeader *fh, FILE *fd, StringTable &table, bela::error_c
   }
   table.length = l;
   return true;
-}
-
-void File::Free() {
-  if (fd != nullptr) {
-    fclose(fd);
-    fd = nullptr;
-  }
-}
-
-void File::FileMove(File &&other) {
-  Free();
-  fd = other.fd;
-  other.fd = nullptr;
-}
-
-std::optional<File> File::NewFile(std::wstring_view p, bela::error_code &ec) {
-  //
-  return std::nullopt;
 }
 } // namespace bela::pe
